# docs/api/openapi.bundled.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Mokhtasr",
		"description": "",
		"version": "1.0.0"
	},
...
	"paths": {
...
		"/register/handyman/{userId}": {
			"post": {
				"summary": "Handyman",
				"deprecated": false,
				"description": "",
				"tags": ["Register"],
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"description": "",
						"required": true,
						"example": "23f89f3a-c0ae-4163-b380-af3e53bde363",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"firstName": {
										"example": "Seif",
										"type": "string"
									},
									"lastName": {
										"example": "Essam",
										"type": "string"
									},
									"nationalIdFront": {
										"format": "binary",
										"type": "string"
									},
									"nationalIdBack": {
										"format": "binary",
										"type": "string"
									},
									"selfie": {
										"format": "binary",
										"type": "string"
									},
									"birthdate": {
										"example": "2001-01-29T21:35:50.599Z",
										"type": "string",
										"format": "date-time"
									},
									"regions": {
										"type": "array",
										"items": {
											"type": "string",
											"format": "uuid"
										},
										"example": ["13a1483c-0950-4666-bd1a-5ea00be0451d", "7cfc85a0-5197-40c2-a112-d13cdda85b1c"]
									},
									"services": {
										"type": "array",
										"items": {
											"type": "string",
											"format": "uuid"
										},
										"example": ["10c3cb25-7a22-4a76-8279-160b111254a9", "514f1e88-8f64-4727-9490-a3867afe8fd7", "0b93d2df-77bb-4e68-8e34-20c0056e078a"]
									}
								},
								"required": ["firstName", "lastName", "nationalIdFront", "nationalIdBack", "selfie", "birthdate", "regions", "services"]
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"handyman": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"format": "uuid"
												},
												"workDays": {
													"type": "array",
													"items": {
														"type": "string",
														"enum": ["SUNDAY", "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY"]
													}
												},
												"status": {
													"type": "string",
													"default": "PENDING",
													"enum": ["PENDING", "REJECTED", "BLOCKED", "ACCEPTED"]
												},
												"birthdate": {
													"type": "string",
													"format": "date-time",
													"nullable": true
												},
												"userId": {
													"type": "string",
													"format": "uuid"
												},
												"user": {
													"$ref": "#/components/schemas/User",
													"description": "Present whenever the handyman is fetched through HandymanService (list/find/update/delete/block)"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												}
											},
											"required": ["id", "workDays", "status", "birthdate", "userId", "createdAt", "updatedAt"]
										},
										"application": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"format": "uuid"
												},
												"firstName": {
													"type": "string"
												},
												"lastName": {
													"type": "string"
												},
												"nationalIdFront": {
													"type": "string",
													"format": "uri"
												},
												"nationalIdBack": {
													"type": "string",
													"format": "uri"
												},
												"selfie": {
													"type": "string",
													"format": "uri"
												},
												"birthdate": {
													"type": "string",
													"format": "date-time"
												},
												"handymanId": {
													"type": "string",
													"format": "uuid"
												},
												"status": {
													"type": "string",
													"default": "PENDING",
													"enum": ["PENDING", "REJECTED", "EXPIRED", "ACCEPTED"]
												},
												"expiresAt": {
													"type": "string",
													"format": "date-time"
												},
												"notes": {
													"type": "string",
													"nullable": true
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"regions": {
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"format": "uuid"
															},
															"name": {
																"type": "string"
															}
														},
														"required": ["id", "name"]
													}
												},
												"services": {
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"format": "uuid"
															},
															"name": {
																"type": "object",
																"properties": {
																	"ar": {
																		"type": "string"
																	},
																	"en": {
																		"type": "string"
																	}
																},
																"required": ["ar", "en"]
															}
														},
														"required": ["id", "name"]
													}
												}
											},
											"required": ["id", "firstName", "lastName", "nationalIdFront", "nationalIdBack", "selfie", "birthdate", "handymanId", "status", "expiresAt", "createdAt", "updatedAt", "regions", "services"]
										}
									},
									"required": ["handyman", "application"]
								}
							}
						},
						"headers": {}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"title": "",
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"details": {
													"type": "string",
													"nullable": true
												},
												"message": {
													"type": "string"
												}
											},
											"required": ["message"]
										}
									},
									"required": ["error"]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/register/customer/{userId}": {
			"post": {
				"summary": "Customer",
				"deprecated": false,
				"description": "",
				"tags": ["Register"],
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"description": "",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"example": "Seif",
										"type": "string"
									},
									"avatar": {
										"format": "binary",
										"type": "string"
									},
									"locale": {
										"type": "string",
										"description": "locale",
										"enum": ["AR", "EN"]
									}
								}
							}
						}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "string",
											"format": "uuid"
										},
										"userId": {
											"type": "string",
											"format": "uuid"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": ["id", "userId", "createdAt", "updatedAt"]
								}
							}
						},
						"headers": {}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"title": "",
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"details": {
													"type": "string",
													"nullable": true
												},
												"message": {
													"type": "string"
												}
											},
											"required": ["message"]
										}
									},
									"required": ["error"]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/register/admin/{userId}": {
			"post": {
				"summary": "Admin",
				"deprecated": false,
				"description": "",
				"tags": ["Register"],
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"description": "",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "string",
											"format": "uuid"
										},
										"userId": {
											"type": "string",
											"format": "uuid"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": ["id", "userId", "createdAt", "updatedAt"]
								}
							}
						},
						"headers": {}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"title": "",
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"details": {
													"type": "string",
													"nullable": true
												},
												"message": {
													"type": "string"
												}
											},
											"required": ["message"]
										}
									},
									"required": ["error"]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string",
						"pattern": "^\\+20(10|11|12|15)\\d{8}$"
					},
					"role": {
						"type": "string",
						"enum": ["ADMIN", "HANDYMAN", "CUSTOMER"]
					},
					"avatar": {
						"type": "string",
						"format": "uri",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"admin": {
						"$ref": "#/components/schemas/Admin",
						"nullable": true
					},
					"customer": {
						"$ref": "#/components/schemas/Customer",
						"nullable": true
					},
					"handyman": {
						"$ref": "#/components/schemas/Handyman",
						"nullable": true
					}
				},
				"required": ["id", "name", "phone", "role", "createdAt", "updatedAt"]
			},
			"Handyman": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"workDays": {
						"type": "array",
						"items": {
							"type": "string",
							"enum": ["SUNDAY", "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY"]
						}
					},
					"status": {
						"type": "string",
						"default": "PENDING",
						"enum": ["PENDING", "REJECTED", "BLOCKED", "ACCEPTED"]
					},
					"birthdate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"userId": {
						"type": "string",
						"format": "uuid"
					},
					"user": {
						"$ref": "#/components/schemas/User",
						"description": "Present whenever the handyman is fetched through HandymanService (list/find/update/delete/block)"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": ["id", "workDays", "status", "birthdate", "userId", "createdAt", "updatedAt"]
			},
...
			"Admin": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"userId": {
						"type": "string",
						"format": "uuid"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": ["id", "userId", "createdAt", "updatedAt"]
			},
			"Customer": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"userId": {
						"type": "string",
						"format": "uuid"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": ["id", "userId", "createdAt", "updatedAt"]
			},
...
		},
...
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Development"
		},
		{
			"url": "https://mokhtasr-api.urax.dev",
			"description": "Staging"
		}
	],
	"security": [
		{
			"bearer": []
		}
	]
}